# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Description of specifications",
        "title": "API document title",
        "contact": {
            "name": "API supporter",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "license(Mandatory)",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "version(1.0)"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/balance": {
            "post": {
                "description": "get all balance from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Show balance.",
                "operationId": "get-all-Balance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "struct"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Accounts"
                        }
                    }
                }
            }
        },
        "/top_up": {
            "post": {
                "description": "Top-up e-wallet account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TopUps"
                ],
                "summary": "Top up the Balance.",
                "operationId": "top-up",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "struct"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.TopUps"
                        }
                    }
                }
            }
        },
        "/total": {
            "post": {
                "description": "Get total count and total amount of top-up operations for current month.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TopUps"
                ],
                "summary": "Get total balance.",
                "operationId": "get-all-Students",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "struct"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.TopUps"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "post": {
                "description": "Check user if it is exist or not.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "Check user.",
                "operationId": "check-acount",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Users"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Users"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Accounts": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identified": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.TopUps": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.Users": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}